save cart auto/ no auto

+++ src/shop/views.py
@@ -1,5 +1,3 @@
-import pprint
-
 from django.contrib.auth import get_user_model
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.http import HttpResponse, HttpResponseRedirect
@@ -11,7 +9,7 @@ from django.views.generic import (CreateView, DetailView, ListView,
 
 from shop.forms import (OrderForm, ProductFilterForm,
                         ProductFilterWithCategoryForm)
-from shop.models import Cart, CartItem, Category, Order, Product
+from shop.models import Cart, CartItem, Category, GuestCart, Order, Product
 from shop.tasks import generate_product_brand_category
 from shop.utils.generate_unique_session_id import generate_unique_session_id
 from shop.utils.sort_queryset_by_price import sort_queryset_by_price
@@ -90,9 +88,22 @@ class ProductDetailView(DetailView):
 class AddToCartView(View):
     def post(self, request, product_id):
         product = Product.objects.get(pk=product_id)
+
         if request.user.is_authenticated:
             customer = request.user
             cart, created = Cart.objects.get_or_create(customer=customer)
+
+            guest_session_id = request.session.get("guest_session_id")
+            if guest_session_id:
+                guest_cart = GuestCart.objects.filter(session_key=guest_session_id).first()
+                if guest_cart:
+                    for guest_cart_item in guest_cart.guestcartitem_set.all():
+                        cart_item, created = CartItem.objects.get_or_create(cart=cart, product=guest_cart_item.product)
+                        if not created:
+                            cart_item.quantity += guest_cart_item.quantity
+                            cart_item.save()
+                    guest_cart.delete()
+                    del request.session["guest_session_id"]
         else:
             guest_session_id = request.session.get("guest_session_id")
             if not guest_session_id:
@@ -100,7 +111,7 @@ class AddToCartView(View):
                 request.session["guest_session_id"] = guest_session_id
             customer = None
 
-            cart, created = Cart.objects.get_or_create(guest_session_id=guest_session_id)
+            cart, created = Cart.objects.get_or_create(customer=None, guest_session_id=guest_session_id)
 
         cart_item, created = CartItem.objects.get_or_create(cart=cart, product=product)
         if not created:
@@ -146,22 +157,42 @@ class UpdateCartView(RedirectView):
     def get_redirect_url(self, *args, **kwargs):
         product_id = kwargs["product_id"]
         quantity = int(self.request.POST.get("quantity"))
-        cart_item = get_object_or_404(CartItem, product_id=product_id, cart__customer=self.request.user)
 
-        if quantity <= 0:
-            cart_item.delete()
+        if self.request.user.is_authenticated:
+            cart_item = get_object_or_404(CartItem, product_id=product_id, cart__customer=self.request.user)
+
+            if quantity <= 0:
+                cart_item.delete()
+            else:
+                cart_item.quantity = quantity
+                cart_item.save()
+
         else:
-            cart_item.quantity = quantity
-            cart_item.save()
+            guest_session_id = self.request.session.get("guest_session_id")
+            if guest_session_id:
+                cart_item = get_object_or_404(CartItem, product_id=product_id, cart__guest_session_id=guest_session_id)
 
-        return reverse("shop:cart")
+                if quantity <= 0:
+                    cart_item.delete()
+                else:
+                    cart_item.quantity = quantity
+                    cart_item.save()
+
+        return self.request.META.get("HTTP_REFERER", reverse("shop:cart"))
 
 
 class RemoveFromCartView(View):
     def post(self, request, product_id):
-        cart_item = get_object_or_404(CartItem, product_id=product_id, cart__customer=request.user)
-        cart_item.delete()
-        return HttpResponseRedirect(reverse("shop:cart"))
+        if request.user.is_authenticated:
+            cart_item = get_object_or_404(CartItem, product_id=product_id, cart__customer=request.user)
+            cart_item.delete()
+        else:
+            guest_session_id = request.session.get("guest_session_id")
+            if guest_session_id:
+                cart_item = get_object_or_404(CartItem, product_id=product_id, cart__guest_session_id=guest_session_id)
+                cart_item.delete()
+
+        return HttpResponseRedirect(request.META.get("HTTP_REFERER", "/"))
 
 
 class CreateOrderView(LoginRequiredMixin, CreateView):
@@ -174,14 +205,11 @@ class CreateOrderView(LoginRequiredMixin, CreateView):
         current_user = self.request.user
         form.instance.customer = current_user
 
-        # Сначала сохраните заказ в базе данных, чтобы получить значение "id"
         response = super().form_valid(form)
 
-        # Затем получите продукты из корзины пользователя и добавьте их в заказ
         cart_items = current_user.cart.cart_items.all()
         self.object.products.set([item.product for item in cart_items])
 
-        # Очистите корзину пользователя
         cart_items.delete()
         return response
